requirements.txt
streamlit
python-dotenv
openai
sentence-transformers
faiss-cpu
PyMuPDF
numpy

